Implement ModifyLandingPage (and supporting code and tests)

+++ test.sh
@@ -1,9 +1,25 @@
 #!/bin/bash
 
-set -e -u
+set -e
 
-rm -rf build
-mkdir build
+CLEAN=
+if [ "$1" == "clean" ] || [ "$2" == "clean" ]; then
+  CLEAN=1
+fi
+COVER=
+if [ "$1" == "cover" ] || [ "$2" == "cover" ]; then
+  COVER=1
+fi
+
+# Don't set this until after checking $1 and $2
+set -u
+
+if [ ${CLEAN} ]; then
+  rm -rf build
+fi
+find . -name "*.gcda" -print0 | xargs -0 rm
+
+mkdir -p build
 cd build
 #export CC=$(which clang) CXX=$(which clang++)
 cmake ..
@@ -12,12 +28,14 @@ cd -
 
 ./build/runUnitTests
 
-gcov -o build/CMakeFiles/psicash.dir/*.gcno *.cpp > /dev/null
+if [ ${COVER} ]; then
 
-lcov --capture --directory . --output-file build/coverage.info > /dev/null
-genhtml build/coverage.info --output-directory build/cov > /dev/null
-echo "Coverage output in $(pwd)/build/cov/index.html"
+  gcov -o build/CMakeFiles/psicash.dir/*.gcno *.cpp > /dev/null
 
-rm *.gcov
+  lcov --capture --directory . --output-file build/coverage.info > /dev/null
+  genhtml build/coverage.info --output-directory build/cov > /dev/null
+  echo "Coverage output in $(pwd)/build/cov/index.html"
 
-#open build/cov/index.html
+  rm *.gcov
+  #open build/cov/index.html
+fi
